rename: piece들을 piece 패키지로 이동
