+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
+
+import chess.domain.Color;
+import chess.domain.Position;
 
 public class Bishop extends Piece {
     public Bishop(Position position, Color color) {
