+++ src/main/java/chess/domain/piece/King.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
+
+import chess.domain.Color;
+import chess.domain.Position;
 
 public class King extends Piece {
     public King(Position position, Color color) {
