+++ src/main/java/chess/domain/piece/Night.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
+
+import chess.domain.Color;
+import chess.domain.Position;
 
 public class Night extends Piece {
     public Night(Position position, Color color) {
