+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
+
+import chess.domain.Color;
+import chess.domain.Position;
 
 public class Pawn extends Piece {
     public Pawn(Position position, Color color) {
