+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
+
+import chess.domain.Color;
+import chess.domain.Position;
 
 public class Queen extends Piece {
     public Queen(Position position, Color color) {
