+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
+
+import chess.domain.Color;
+import chess.domain.Position;
 
 public class Rook extends Piece {
     public Rook(Position position, Color color) {
