+++ src/test/java/chess/domain/ChessGameTest.java
@@ -10,6 +10,14 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import chess.domain.piece.Bishop;
+import chess.domain.piece.King;
+import chess.domain.piece.Night;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
+
 public class ChessGameTest {
     @ParameterizedTest(name = "{index}: {2}")
     @MethodSource("invalidParameters")
